+++ huey/__init__.py
@@ -1,6 +1,6 @@
 __author__ = 'Charles Leifer'
 __license__ = 'MIT'
-__version__ = '3.2.0'
+__version__ = '3.2.1'
 
 from huey.api import BlackHoleHuey
 from huey.api import Huey